50~60 %
